+++ Native2Forms/Notes.iOS/Views/NoteDetailsPage.xaml.cs
@@ -1,10 +1,12 @@
 ﻿using System;
-using System.Collections.Generic;
-
+using System.IO;
+using Notes.iOS.Models;
 using Xamarin.Forms;
+using Xamarin.Forms.Xaml;
 
-namespace Notes.Views
+namespace Notes.iOS.Views
 {
+    [XamlCompilation(XamlCompilationOptions.Compile)]
     public partial class NoteDetailsPage : ContentPage
     {
         public NoteDetailsPage()
@@ -16,14 +18,26 @@ namespace Notes.Views
         {
             base.OnAppearing();
 
-            ViewHolder.TranslateTo(0, 0, 600, Easing.BounceIn);
+            grid.TranslateTo(0, 0, 600, Easing.BounceIn);
         }
 
         protected override void OnDisappearing()
         {
             base.OnDisappearing();
             
-            ViewHolder.TranslateTo(0, -2000, 300, Easing.BounceOut);
+            grid.TranslateTo(0, -2000, 300, Easing.BounceOut);
+        }
+
+        void OnDeleteButtonClicked(object sender, EventArgs e)
+        {
+            var note = (Note)BindingContext;
+
+            if (File.Exists(note.Filename))
+            {
+                File.Delete(note.Filename);
+            }
+
+            AppDelegate.Instance.NavigateBack();
         }
     }
 }
