+++ Native2Forms/Notes.iOS/Views/NoteEntryPage.xaml
@@ -6,16 +6,8 @@
         <Editor Placeholder="Enter your note"
                 Text="{Binding Text}"
                 HeightRequest="100" />
-        <Grid>
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="*" />
-                <ColumnDefinition Width="*" />
-            </Grid.ColumnDefinitions>
-            <Button Text="Save" 
-                    Clicked="OnSaveButtonClicked" />
-            <Button Grid.Column="1"
-                    Text="Delete" 
-                    Clicked="OnDeleteButtonClicked"/>
-        </Grid>
+        <Button Text="Save"
+                HorizontalOptions="Center"
+                Clicked="OnSaveButtonClicked" />
     </StackLayout>
 </ContentPage>
